requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
Gunicorn>=20.0.0
python-dotenv>=1.0.0
google-adk
selenium
undetected_chromedriver